+++ packages/mobile-native-mcp-server/src/common/schemas/common.ts
