--- packages/mobile-native-mcp-server/src/logging/index.ts
@@ -1,20 +1,0 @@
-/**
- * Logging module exports
- *
- * Centralized logging functionality for the Mobile Native MCP Server.
- * Provides consistent logging across all tools and facilities.
- */
-
-export {
-  Logger,
-  PinoLogger,
-  createProductionLogger,
-  createDevelopmentLogger,
-  createTestLogger,
-  createLogger,
-  createComponentLogger,
-  createWorkflowLogger,
-} from './logger.js';
-
-// Re-export commonly used utilities for convenience
-export { createComponentLogger as createToolLogger } from './logger.js';
