--- packages/mobile-native-mcp-server/src/schemas/index.ts
@@ -1,20 +1,0 @@
-/**
- * Schema Module Exports
- *
- * Centralized exports for all schema definitions. This acts as a barrel file
- * to provide clean imports throughout the codebase.
- */
-
-// Common schema components
-export * from './common.js';
-
-// Workflow-specific schemas
-export * from './workflow.js';
-
-// Tool schemas
-export * from './tools/orchestrator.js';
-export * from './tools/template-discovery.js';
-export * from './tools/project-generation.js';
-export * from './tools/build.js';
-export * from './tools/deployment.js';
-export * from './tools/xcode-add-files.js';
